+++ index.css
@@ -25,7 +25,6 @@
   margin: 5px;
   padding: 5px;
   width: 200px;
-  height: 300px;
 }
 .res-card:hover {
   cursor: pointer;
