+++ src/App.js
@@ -1,0 +1,15 @@
+import ReactDOM from "react-dom/client";
+import Header from "./components/Header.js";
+import Body from "./components/Body.js";
+
+const AppLayout = () => {
+  return (
+    <div className="app">
+      <Header />
+      <Body />
+    </div>
+  );
+};
+
+const root = ReactDOM.createRoot(document.getElementById("root"));
+root.render(<AppLayout />);
