+++ src/components/Header.js
@@ -1,0 +1,18 @@
+import { LOGO_URL } from "../utils/constants";
+const Header = () => (
+  <div className="header">
+    <div className="logoContainer">
+      <img className="logo" src={LOGO_URL}></img>
+    </div>
+    <div className="nav-items">
+      <ul>
+        <li>Home </li>
+        <li>About Us</li>
+        <li>Contact Us </li>
+        <li>Cart</li>
+      </ul>
+    </div>
+  </div>
+);
+
+export default Header;
