+++ src/utils/mockData.js
@@ -1,0 +1,37 @@
+export const resList = [
+  {
+    id: 1,
+    name: "Spicy Village",
+    cuisine: "Chinese",
+    rating: 4.5,
+    image: "https://picsum.photos/200?random=1",
+  },
+  {
+    id: 2,
+    name: "The Italian Bistro",
+    cuisine: "Italian",
+    rating: 4.7,
+    image: "https://picsum.photos/200?random=2",
+  },
+  {
+    id: 3,
+    name: "Taco Fiesta",
+    cuisine: "Mexican",
+    rating: 4.2,
+    image: "https://picsum.photos/200?random=3",
+  },
+  {
+    id: 4,
+    name: "Sushi Zen",
+    cuisine: "Japanese",
+    rating: 4.6,
+    image: "https://picsum.photos/200?random=4",
+  },
+  {
+    id: 5,
+    name: "Café Bonjour",
+    cuisine: "French",
+    rating: 4.8,
+    image: "https://picsum.photos/200?random=5",
+  },
+];
