+++ src/utils/styleCard.js
@@ -1,0 +1,4 @@
+const styleCard = {
+  backgroundColor: "#f0f0f0",
+};
+export default styleCard;
